# docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/v1/leases:
    get:
      consumes:
      - application/json
      description: Retrieves all leases in the specified namespace
      parameters:
      - description: Kubernetes config ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name
        in: query
        name: cluster
        type: string
      - description: Namespace name
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of leases
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get all leases in a namespace
      tags:
      - Leases
  /api/v1/leases/{name}/events:
    get:
      consumes:
      - application/json
      description: Retrieves events related to a specific lease
      parameters:
      - description: Kubernetes config ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name
        in: query
        name: cluster
        type: string
      - description: Lease name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      - text/event-stream
      responses:
        "200":
          description: List of events
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get lease events
      tags:
      - Leases
  /api/v1/leases/{name}/yaml:
    get:
      consumes:
      - application/json
      description: Retrieves the YAML representation of a specific lease
      parameters:
      - description: Kubernetes config ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name
        in: query
        name: cluster
        type: string
      - description: Namespace name
        in: query
        name: namespace
        required: true
        type: string
      - description: Lease name
        in: path
        name: name
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Lease YAML
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Lease not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get lease YAML
      tags:
      - Leases
  /api/v1/leases/{namespace}/{name}:
    get:
      consumes:
      - application/json
      description: Retrieves a specific lease by name and namespace
      parameters:
      - description: Kubernetes config ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name
        in: query
        name: cluster
        type: string
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      - description: Lease name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      - text/event-stream
      responses:
        "200":
          description: Lease details
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Lease not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get a specific lease
      tags:
      - Leases
  /api/v1/leases/sse:
    get:
      consumes:
      - application/json
      description: Retrieves leases in the specified namespace with Server-Sent Events
        for real-time updates
      parameters:
      - description: Kubernetes config ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name
        in: query
        name: cluster
        type: string
      - description: Namespace name
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - text/event-stream
      - application/json
      responses:
        "200":
          description: Stream of leases or JSON array
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get leases with real-time updates
      tags:
      - Leases
...
swagger: "2.0"